Starting Point
Not sure what happened. this is when I thought I lost the project, so I just opened the project from the GitHub. Might be the main and then all of these changes in the branch, idk, but this is the starting point for this branch. Most importantly, it all works.