almostCompleted
Scaled graph pretty much completed. A little sketchy, and a small bug where in month 25, the line goes off of the graph, but other than that it is great. Also, the start of the graph is at $1,000 and one month, which is why it looks so off on the first iteration.

Things to add:
- starting line at origin
- add text indicating highest value represented on the graph
- fix error of month 25 going off of the grid