Generated MoreInfo UI and Budgeting list of links
Generated the menu with the tab screen and scroll screen for the more info button that will house the links to all of the videos on AloraFinance.com.
Built the scroll screen for the Budgeting topics. Need to do it for the other 4 and also hook up all of the code to make it work