Working everything but with a couple bugs
The queue is working, every order has a number and the delivery button works. Still some bugs though (can't figure out the assignment of all the public variables in the inspector of both scripts attached to the canvas).
Also, the current order prefab is spawning off the screen for some reason. I don't think it needs to spawn off of the screen, just update the text of the existing one every time there is a new order to complete.